feat: implement NativeAssetMap
